chore: refactor statemachine tests
