+++ src/server/handlers/participants/{Type}/{ID}.ts
@@ -23,8 +23,8 @@
  ******/
 import { Request, ResponseObject } from '@hapi/hapi'
 import { Enum } from '@mojaloop/central-services-shared'
-import { RegisterConsentPhase } from '~/model/registerConsent.interface'
-import { RegisterConsentModel } from '~/model/registerConsent.model'
+import { RegisterConsentPhase } from '~/domain/stateMachine/registerConsent.interface'
+import { RegisterConsentModel } from '~/domain/stateMachine/registerConsent.model'
 import { thirdparty as tpAPI } from '@mojaloop/api-snippets'
 import { StateResponseToolkit } from '~/server/plugins/state'
 import { Message } from '~/shared/pub-sub'
